refactor: merge multiple EntryPointType structs
