+++ crates/blockifier/src/transaction/transactions_test.rs
@@ -9,8 +9,8 @@ use num_traits::Pow;
 use pretty_assertions::assert_eq;
 use rstest::{fixture, rstest};
 use starknet_api::block::GasPriceVector;
+use starknet_api::contract_class::EntryPointType;
 use starknet_api::core::{ChainId, ClassHash, ContractAddress, EthAddress, Nonce, PatriciaKey};
-use starknet_api::deprecated_contract_class::EntryPointType;
 use starknet_api::execution_resources::GasVector;
 use starknet_api::state::StorageKey;
 use starknet_api::test_utils::invoke::InvokeTxArgs;
