+++ crates/papyrus_rpc/src/v0_8/deprecated_contract_class.rs
@@ -2,7 +2,8 @@ use std::collections::HashMap;
 
 use papyrus_storage::db::serialization::StorageSerdeError;
 use serde::{Deserialize, Serialize};
-use starknet_api::deprecated_contract_class::{ContractClassAbiEntry, EntryPoint, EntryPointType};
+use starknet_api::contract_class::EntryPointType;
+use starknet_api::deprecated_contract_class::{ContractClassAbiEntry, EntryPoint};
 
 use crate::compression_utils::compress_and_encode;
