+++ crates/starknet_client/src/writer/objects/test_utils.rs
@@ -1,11 +1,11 @@
 use std::collections::HashMap;
 
 use papyrus_test_utils::{auto_impl_get_test_instance, get_number_of_variants, GetTestInstance};
+use starknet_api::contract_class::EntryPointType;
 use starknet_api::core::{ClassHash, ContractAddress};
 use starknet_api::deprecated_contract_class::{
     ContractClassAbiEntry as DeprecatedContractClassAbiEntry,
     EntryPoint as DeprecatedEntryPoint,
-    EntryPointType as DeprecatedEntryPointType,
 };
 use starknet_api::transaction::TransactionHash;
 
@@ -21,7 +21,7 @@ auto_impl_get_test_instance! {
     pub struct DeprecatedContractClass {
         pub abi: Option<Vec<DeprecatedContractClassAbiEntry>>,
         pub compressed_program: String,
-        pub entry_points_by_type: HashMap<DeprecatedEntryPointType, Vec<DeprecatedEntryPoint>>,
+        pub entry_points_by_type: HashMap<EntryPointType, Vec<DeprecatedEntryPoint>>,
     }
     pub struct InvokeResponse {
         pub code: SuccessfulStarknetErrorCode,
